Updated Data Collection Process
I started trying to actually track the player data using unity analytics and found that the initial process was not working. I got rid of DataConsent and MainMenu and now just have one class for GameSession.
