--- Demo/Assets/Exitgame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9b0443e0533dc44b9b5eed6234dfb3f8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
