+++ Demo/Assets/GameSession.cs.meta
@@ -1,9 +1,5 @@
 fileFormatVersion: 2
-<<<<<<< Updated upstream
-guid: 15c8b914e02c1054bbe6a1bb6d671285
-=======
-guid: b58a4999b01014c5595b3f8e2402fbfe
->>>>>>> Stashed changes
+guid: 973316e968a344c8ab09a253748601a4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
