--- Demo/Assets/MainMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b839116af698f40a992270d254761dac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
