+++ Demo/Assets/Player.cs
@@ -1,6 +1,8 @@
 using System.Collections;
 using System.Collections.Generic;
+using Unity.VisualScripting;
 using UnityEngine;
+using UnityEngine.SceneManagement;
 
 public class Player : Actor
 {
@@ -24,14 +26,17 @@ public class Player : Actor
     float requiredXP = 45;
     public float currentXP;
 
+
     // Start is called before the first frame update
     private void Start()
     {
+        Debug.Log("HERE");
         rb = GetComponent<Rigidbody2D>();
         walkForce = 9;
         jumpForce = 15;
         SetMaxHP(30);
         AdjustXP(0);
+
     }
 
     private void UpdateMovement()
